fix: restore model types deleted in 227dbe6, fix NFT collection page
- Restore ~150 lines of model definitions (DefiEvent, DefiPair,
  ContractVersion, StakingEvent, StakingNode, StakingDelegator,
  EpochStats, NodeMetadata, NFTItem, expanded FTToken/NFTCollection)
  that were accidentally removed during the COA badge commit
- Fix NFT collection holder count: add COUNT(DISTINCT owner) to
  GetNFTCollectionSummary query
- Add NFT items fallback from nft_ownership when nft_items is empty
- Add normalizeEVMAddress helper for EVM address normalization
- Add EnsureDefiPartitions/EnsureStakingPartitions no-op stubs
- Fix uint64/int64 type mismatches in staking_worker
- Remove unnecessary dependency gate on metadata worker

+++ backend/internal/ingester/staking_worker.go
@@ -178,8 +178,8 @@ func (w *StakingWorker) processNodeEvent(
 	}
 
 	// Always update first_seen_height
-	if node.FirstSeenHeight == 0 || int64(evt.BlockHeight) < node.FirstSeenHeight {
-		node.FirstSeenHeight = int64(evt.BlockHeight)
+	if node.FirstSeenHeight == 0 || evt.BlockHeight < node.FirstSeenHeight {
+		node.FirstSeenHeight = evt.BlockHeight
 	}
 
 	switch eventName {
@@ -248,7 +248,7 @@ func (w *StakingWorker) processEpochEvent(
 
 		// Use the event's block height as epoch start height.
 		// Note: firstView is a consensus view number, NOT a block height.
-		startHeight := int64(evt.BlockHeight)
+		startHeight := evt.BlockHeight
 
 		return &models.EpochStats{
 			Epoch:       epoch,
